live-stream-analyzer/requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
yt-dlp>=2024.3.10
pillow>=10.4.0
opencv-python>=4.8.0
